refactor(shell): extract escape_state into pub module
Move EscapeState, process_escape_events, and cleanup_forwarded_modes out
of impl ShellSession and into devenv-shell/src/escape_state.rs as free
items, plus promote escape and terminal_commands to pub modules. Pure
mechanical refactor preserving session.rs behaviour byte for byte;
the EscapeState unit tests move with the type.

Enables devenv-reload to reuse the same mode tracking and on-exit reset
without duplicating ~250 lines of escape parsing.
