+++ devenv-shell/src/session.rs
@@ -3,14 +3,17 @@
 //! This module provides the main `ShellSession` type that orchestrates
 //! PTY lifecycle, terminal I/O, and status line rendering.
 
-use crate::escape::{CLEANUP_MODES, DecModeEvent, EscapeScanner, SequenceEvent};
+use crate::escape::EscapeScanner;
+use crate::escape_state::{
+    EscapeState, cleanup_forwarded_modes as escape_state_cleanup,
+    process_escape_events as escape_state_process,
+};
 use crate::protocol::{ShellCommand, ShellEvent};
 use crate::pty::{Pty, PtyError, get_terminal_size};
 use crate::status_line::{SPINNER_INTERVAL_MS, StatusLine};
 use crate::terminal::RawModeGuard;
 use crate::terminal_commands::{
-    InBandResizeNotification, ORIGIN_MODE, ReportTextAreaSize, ResetDecMode, ResetModifyOtherKeys,
-    ResetScrollRegion, SetKeypadMode, SetScrollRegion,
+    InBandResizeNotification, ORIGIN_MODE, ResetDecMode, ResetScrollRegion, SetScrollRegion,
 };
 use crate::utf8_accumulator::Utf8Accumulator;
 use crate::vt_utils::{
@@ -18,9 +21,7 @@ use crate::vt_utils::{
     screen_point,
 };
 use crossterm::{
-    Command, cursor,
-    event::PopKeyboardEnhancementFlags,
-    queue,
+    Command, cursor, queue,
     style::ResetColor,
     terminal::{self, Clear, ClearType},
 };
@@ -28,7 +29,6 @@ use libghostty_vt::screen::{Cell, CellWide};
 use libghostty_vt::style::{Style, StyleColor, Underline};
 use libghostty_vt::terminal::{Options as TerminalOptions, Point, PointCoordinate, Terminal};
 use portable_pty::PtySize;
-use std::collections::BTreeSet;
 use std::fmt::Write as FmtWrite;
 use std::io::{self, IsTerminal, Read, Write};
 use std::sync::Arc;
@@ -40,108 +40,6 @@ use tokio::sync::{Notify, mpsc as tokio_mpsc, oneshot};
 const KEYBIND_TOGGLE_PAUSE: [u8; 2] = [0x1b, 0x04]; // Ctrl-Alt-D
 const KEYBIND_LIST_WATCHED: [u8; 2] = [0x1b, 0x17]; // Ctrl-Alt-W
 const KEYBIND_TOGGLE_ERROR: [u8; 2] = [0x1b, 0x05]; // Ctrl-Alt-E
-
-/// Escape-sequence state tracked across PTY output processing.
-///
-/// Persistent fields (`in_alternate_screen`, `forwarded_dec_modes`,
-/// `keypad_application_mode`) carry across the entire session.
-/// Per-batch fields (`erase_display`, `clear_scrollback`) are reset at the
-/// start of each `PtyOutput` batch.
-struct EscapeState {
-    in_alternate_screen: bool,
-    /// DEC private modes that were set and need explicit reset on exit.
-    /// Tracked separately from `in_alternate_screen` (which uses
-    /// `LeaveAlternateScreen`) and keypad mode (which uses `ESC >`).
-    forwarded_dec_modes: BTreeSet<u16>,
-    /// Keypad is in application mode (DECKPAM, `ESC =`).
-    keypad_application_mode: bool,
-    /// Set when CSI 2 J is seen — signals the caller to consume `row_offset`.
-    erase_display: bool,
-    /// Set when CSI 3 J is seen — deferred so the caller can emit it *after*
-    /// `scroll_region` pushes old TUI content into scrollback.
-    clear_scrollback: bool,
-    /// Kitty keyboard protocol stack depth.
-    kitty_keyboard_depth: u32,
-    /// XTMODIFYOTHERKEYS is enabled.
-    modify_other_keys: bool,
-    /// Mode 2048 (in-band resize) is enabled by the PTY program.
-    in_band_resize: bool,
-}
-
-impl EscapeState {
-    fn new() -> Self {
-        Self {
-            in_alternate_screen: false,
-            forwarded_dec_modes: BTreeSet::new(),
-            keypad_application_mode: false,
-            erase_display: false,
-            clear_scrollback: false,
-            kitty_keyboard_depth: 0,
-            modify_other_keys: false,
-            in_band_resize: false,
-        }
-    }
-
-    /// Reset per-batch flags before processing a new `PtyOutput` batch.
-    fn reset_batch(&mut self) {
-        self.erase_display = false;
-        self.clear_scrollback = false;
-    }
-
-    /// Apply a DEC mode event, updating tracked state.
-    /// Returns the raw bytes to forward to stdout (empty if mode isn't forwarded).
-    fn apply_dec_mode<'a>(&mut self, event: &'a DecModeEvent) -> &'a [u8] {
-        if event.enables_in_band_resize() {
-            self.in_band_resize = true;
-        } else if event.disables_in_band_resize() {
-            self.in_band_resize = false;
-        }
-
-        if !event.has_forwarded_mode() {
-            return &[];
-        }
-
-        if event.enters_alt_screen() {
-            self.in_alternate_screen = true;
-        } else if event.exits_alt_screen() {
-            self.in_alternate_screen = false;
-        }
-
-        match event {
-            DecModeEvent::Set { modes, .. } => {
-                for &m in modes {
-                    if CLEANUP_MODES.contains(&m) {
-                        self.forwarded_dec_modes.insert(m);
-                    }
-                }
-            }
-            DecModeEvent::Reset { modes, .. } => {
-                for m in modes {
-                    self.forwarded_dec_modes.remove(m);
-                }
-            }
-        }
-
-        event.raw_bytes()
-    }
-
-    /// Apply a kitty keyboard push/pop/set event.
-    fn apply_kitty_keyboard(&mut self, stack_delta: i8) {
-        if stack_delta > 0 {
-            self.kitty_keyboard_depth =
-                self.kitty_keyboard_depth.saturating_add(stack_delta as u32);
-        } else if stack_delta < 0 {
-            self.kitty_keyboard_depth = self
-                .kitty_keyboard_depth
-                .saturating_sub(stack_delta.unsigned_abs() as u32);
-        }
-    }
-
-    /// Apply an XTMODIFYOTHERKEYS set/reset event.
-    fn apply_modify_other_keys(&mut self, enabled: bool) {
-        self.modify_other_keys = enabled;
-    }
-}
 
 /// Render a VT row as a string with SGR escape sequences.
 ///
@@ -1035,7 +933,7 @@ impl ShellSession {
                 Event::PtyOutput(data) => {
                     let was_in_alt = esc.in_alternate_screen;
                     esc.reset_batch();
-                    Self::process_escape_events(
+                    escape_state_process(
                         &mut scanner,
                         &data,
                         &mut esc,
@@ -1053,7 +951,7 @@ impl ShellSession {
                     while let Ok(event) = event_rx.try_recv() {
                         match event {
                             Event::PtyOutput(more) => {
-                                Self::process_escape_events(
+                                escape_state_process(
                                     &mut scanner,
                                     &more,
                                     &mut esc,
@@ -1066,7 +964,7 @@ impl ShellSession {
                                 total_scroll += feed_vt(vt, &text);
                             }
                             Event::PtyExit(exit_code) => {
-                                Self::cleanup_forwarded_modes(&esc, stdout)?;
+                                escape_state_cleanup(&esc, stdout)?;
                                 renderer.render_with_scroll(stdout, vt)?;
                                 return Ok(exit_code);
                             }
@@ -1143,7 +1041,7 @@ impl ShellSession {
 
                 Event::PtyExit(exit_code) => {
                     self.clear_status_row(stdout, esc.in_alternate_screen)?;
-                    Self::cleanup_forwarded_modes(&esc, stdout)?;
+                    escape_state_cleanup(&esc, stdout)?;
                     stdout.flush()?;
                     return Ok(exit_code);
                 }
@@ -1212,7 +1110,7 @@ impl ShellSession {
         }
 
         self.clear_status_row(stdout, esc.in_alternate_screen)?;
-        Self::cleanup_forwarded_modes(&esc, stdout)?;
+        escape_state_cleanup(&esc, stdout)?;
         stdout.flush()?;
         Ok(None)
     }
@@ -1279,111 +1177,6 @@ impl ShellSession {
         Ok(0)
     }
 
-    /// Scan raw PTY output for escape sequences (DEC private mode and OSC queries),
-    /// forward relevant ones to the real terminal, and update escape state.
-    fn process_escape_events(
-        scanner: &mut EscapeScanner,
-        data: &[u8],
-        esc: &mut EscapeState,
-        stdout: &mut impl Write,
-        pty: &Pty,
-        pty_size: PtySize,
-        events_buf: &mut Vec<SequenceEvent>,
-    ) -> io::Result<()> {
-        events_buf.clear();
-        scanner.scan_into(data, events_buf);
-        for event in events_buf.drain(..) {
-            match event {
-                SequenceEvent::DecMode(event) => {
-                    let forward = esc.apply_dec_mode(&event);
-                    if !forward.is_empty() {
-                        stdout.write_all(forward)?;
-                    }
-                }
-                SequenceEvent::Osc(event) => {
-                    // Forward OSC queries to the real terminal so programs can detect color scheme, etc.
-                    stdout.write_all(&event.raw_bytes)?;
-                }
-                SequenceEvent::EraseDisplay { .. } => {
-                    // Not forwarded (the renderer handles screen content), but signals the caller to consume row_offset.
-                    esc.erase_display = true;
-                }
-                SequenceEvent::ClearScrollback { .. } => {
-                    // Deferred so the caller can emit it after scroll_region pushes old TUI content into scrollback.
-                    esc.clear_scrollback = true;
-                }
-                SequenceEvent::ForwardCsi { raw_bytes } => {
-                    stdout.write_all(&raw_bytes)?;
-                }
-                SequenceEvent::ForwardDcs { raw_bytes } => {
-                    stdout.write_all(&raw_bytes)?;
-                }
-                SequenceEvent::KittyKeyboard {
-                    raw_bytes,
-                    stack_delta,
-                } => {
-                    stdout.write_all(&raw_bytes)?;
-                    esc.apply_kitty_keyboard(stack_delta);
-                }
-                SequenceEvent::ModifyOtherKeys { raw_bytes, enabled } => {
-                    stdout.write_all(&raw_bytes)?;
-                    esc.apply_modify_other_keys(enabled);
-                }
-                SequenceEvent::TextAreaSizeQuery => {
-                    // Respond with PTY dimensions so programs see the correct
-                    // size (which excludes the status line row).
-                    let cmd = ReportTextAreaSize {
-                        rows: pty_size.rows,
-                        cols: pty_size.cols,
-                    };
-                    let mut buf = String::new();
-                    cmd.write_ansi(&mut buf).unwrap();
-                    pty.write_all(buf.as_bytes())?;
-                    pty.flush()?;
-                }
-                SequenceEvent::KeypadMode { application } => {
-                    queue!(stdout, SetKeypadMode { application })?;
-                    esc.keypad_application_mode = application;
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Reset any forwarded DEC modes on exit so the terminal is left clean.
-    ///
-    /// XTSHIFTESCAPE and DECSCUSR are forwarded without explicit cleanup:
-    /// mouse tracking modes (cleaned up above) make XTSHIFTESCAPE inert,
-    /// and most terminals reset cursor shape on their own.
-    fn cleanup_forwarded_modes(esc: &EscapeState, stdout: &mut impl Write) -> io::Result<()> {
-        let mut needs_flush = false;
-        if esc.in_alternate_screen {
-            queue!(stdout, terminal::LeaveAlternateScreen)?;
-            needs_flush = true;
-        }
-        for &mode in &esc.forwarded_dec_modes {
-            queue!(stdout, ResetDecMode(mode))?;
-            needs_flush = true;
-        }
-        if esc.keypad_application_mode {
-            queue!(stdout, SetKeypadMode { application: false })?;
-            needs_flush = true;
-        }
-        // Pop all kitty keyboard stack entries
-        for _ in 0..esc.kitty_keyboard_depth {
-            queue!(stdout, PopKeyboardEnhancementFlags)?;
-            needs_flush = true;
-        }
-        if esc.modify_other_keys {
-            queue!(stdout, ResetModifyOtherKeys)?;
-            needs_flush = true;
-        }
-        if needs_flush {
-            stdout.flush()?;
-        }
-        Ok(())
-    }
-
     /// Draw status line and reposition cursor.
     ///
     /// Does not flush — callers flush after ending their sync block.
@@ -1424,130 +1217,5 @@ impl ShellSession {
 impl Default for ShellSession {
     fn default() -> Self {
         Self::with_defaults()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Feed raw bytes through scanner and apply events to state.
-    /// Returns bytes that would be forwarded to stdout.
-    fn scan_and_apply(scanner: &mut EscapeScanner, esc: &mut EscapeState, input: &[u8]) -> Vec<u8> {
-        let events = scanner.scan(input);
-        let mut forwarded = Vec::new();
-        for event in &events {
-            match event {
-                SequenceEvent::DecMode(ev) => forwarded.extend_from_slice(esc.apply_dec_mode(ev)),
-                SequenceEvent::KittyKeyboard { stack_delta, .. } => {
-                    esc.apply_kitty_keyboard(*stack_delta);
-                }
-                SequenceEvent::ModifyOtherKeys { enabled, .. } => {
-                    esc.apply_modify_other_keys(*enabled);
-                }
-                _ => {}
-            }
-        }
-        forwarded
-    }
-
-    #[test]
-    fn in_band_resize_tracked_on_opt_in() {
-        let mut scanner = EscapeScanner::new();
-        let mut esc = EscapeState::new();
-        let forwarded = scan_and_apply(&mut scanner, &mut esc, b"\x1b[?2048h");
-        assert!(esc.in_band_resize);
-        assert!(forwarded.is_empty(), "mode 2048 should not be forwarded");
-    }
-
-    #[test]
-    fn in_band_resize_cleared_on_opt_out() {
-        let mut scanner = EscapeScanner::new();
-        let mut esc = EscapeState::new();
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[?2048h");
-        assert!(esc.in_band_resize);
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[?2048l");
-        assert!(!esc.in_band_resize);
-    }
-
-    #[test]
-    fn in_band_resize_compound_mode() {
-        let mut scanner = EscapeScanner::new();
-        let mut esc = EscapeState::new();
-        // Mode 1 (DECCKM) is forwarded, mode 2048 is not
-        let forwarded = scan_and_apply(&mut scanner, &mut esc, b"\x1b[?1;2048h");
-        assert!(esc.in_band_resize);
-        assert!(!forwarded.is_empty(), "mode 1 should be forwarded");
-    }
-
-    #[test]
-    fn alt_screen_enter_exit() {
-        let mut scanner = EscapeScanner::new();
-        let mut esc = EscapeState::new();
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[?1049h");
-        assert!(esc.in_alternate_screen);
-        // Alt screen modes are not in CLEANUP_MODES (cleaned up via LeaveAlternateScreen)
-
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[?1049l");
-        assert!(!esc.in_alternate_screen);
-    }
-
-    #[test]
-    fn alt_screen_all_variants() {
-        for mode in [47, 1047, 1049] {
-            let mut scanner = EscapeScanner::new();
-            let mut esc = EscapeState::new();
-            let seq = format!("\x1b[?{}h", mode);
-            scan_and_apply(&mut scanner, &mut esc, seq.as_bytes());
-            assert!(
-                esc.in_alternate_screen,
-                "mode {} should enter alt screen",
-                mode
-            );
-        }
-    }
-
-    #[test]
-    fn forwarded_modes_accumulate() {
-        let mut scanner = EscapeScanner::new();
-        let mut esc = EscapeState::new();
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[?1000h");
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[?2004h");
-        assert!(esc.forwarded_dec_modes.contains(&1000));
-        assert!(esc.forwarded_dec_modes.contains(&2004));
-    }
-
-    #[test]
-    fn non_forwarded_mode_no_stdout() {
-        let mut scanner = EscapeScanner::new();
-        let mut esc = EscapeState::new();
-        let forwarded = scan_and_apply(&mut scanner, &mut esc, b"\x1b[?2048h");
-        assert!(forwarded.is_empty());
-    }
-
-    #[test]
-    fn kitty_keyboard_depth_tracking() {
-        let mut scanner = EscapeScanner::new();
-        let mut esc = EscapeState::new();
-        // Push (CSI > u)
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[>1u");
-        assert_eq!(esc.kitty_keyboard_depth, 1);
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[>1u");
-        assert_eq!(esc.kitty_keyboard_depth, 2);
-        // Pop (CSI < u)
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[<u");
-        assert_eq!(esc.kitty_keyboard_depth, 1);
-    }
-
-    #[test]
-    fn modify_other_keys_tracking() {
-        let mut scanner = EscapeScanner::new();
-        let mut esc = EscapeState::new();
-        // CSI > 4 m — enable
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[>4m");
-        assert!(esc.modify_other_keys);
-        // CSI > 0 m — disable
-        scan_and_apply(&mut scanner, &mut esc, b"\x1b[>0m");
-        assert!(!esc.modify_other_keys);
     }
 }
